Fix for firefox resizing and added two additional options
Later versions of Firefox don't have an iframe's body content scrollHeight set to the actual height of the content, but of the scroll height of the element if scroll is disabled.  Using contentDocument.documentElement.scrollHeight gives the correct height more consistently.

I also added an option to reset the iframe to the specified minHeight prior to resizing due to Firefox 14 and IE 9 not resizing "down" if the iframe size was taller than the scrollHeight.
I also added an option to do continual iframe resizes on browser resizing.

Both of the new options are defaulted to false.  I did need those features though, so I thought it might be useful for someone else.