Bump version to 2.0.0
There is very little change in the actual code between 1.9.x and 2.0.x.
The primary changes were in the re-organization of the repository and
introduction of modern javascript build tools (grunt) in place of
hand-rolled ruby scripts.

The demo code is no longer a rails application and instead uses static
html files.